separated methods in new file
